+++ notifications/src/main/java/org/opennms/horizon/notifications/tenant/TenantAwareKafkaListenerAspect.java
@@ -26,12 +26,11 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.notifications.kafka;
+package org.opennms.horizon.notifications.tenant;
 
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
-import org.opennms.horizon.notifications.tenant.TenantContext;
 import org.opennms.horizon.shared.constants.GrpcConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
