--- notifications/src/main/java/org/opennms/horizon/notifications/tenant/TenantAwareTaskDecorator.java
@@ -1,52 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2023 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2023 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.horizon.notifications.tenant;
-
-import org.springframework.core.task.TaskDecorator;
-import org.springframework.lang.NonNull;
-
-public class TenantAwareTaskDecorator implements TaskDecorator {
-
-    @Override
-    @NonNull
-    public Runnable decorate(@NonNull Runnable runnable) {
-        String tenantId = TenantContext.getTenantId();
-        return () -> {
-            try {
-                System.out.println("JH Decorator set:"+tenantId);
-                TenantContext.setTenantId(tenantId);
-                runnable.run();
-            } finally {
-                System.out.println("JH Decorator clear");
-                TenantContext.clear();
-                System.out.println("JH Decorator after clear:"+TenantContext.getTenantId());
-            }
-        };
-    }
-}
