+++ notifications/src/test/java/org/opennms/horizon/notifications/kafka/AlarmKafkaConsumerIntegrationTest.java
@@ -16,14 +16,19 @@ import org.opennms.horizon.notifications.NotificationsApplication;
 import org.opennms.horizon.notifications.SpringContextTestInitializer;
 import org.opennms.horizon.notifications.exceptions.NotificationException;
 import org.opennms.horizon.notifications.service.NotificationService;
+import org.opennms.horizon.notifications.tenant.TenantContext;
+import org.opennms.horizon.notifications.tenant.WithTenant;
 import org.opennms.horizon.shared.constants.GrpcConstants;
 import org.opennms.horizon.shared.dto.event.AlarmDTO;
 import org.opennms.horizon.notifications.dto.PagerDutyConfigDTO;
+import org.springframework.aop.aspectj.annotation.AnnotationAwareAspectJAutoProxyCreator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.mock.mockito.SpyBean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.EnableAspectJAutoProxy;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpMethod;
 import org.springframework.kafka.core.DefaultKafkaProducerFactory;
@@ -89,16 +94,19 @@ class AlarmKafkaConsumerIntegrationTest {
         stringProducer = stringFactory.createProducer();
     }
 
-    private void setupConfig() throws NotificationException {
-        String integrationKey = "not_verified";
+    public void setupConfig(String tenantId) throws NotificationException {
+        // WithTenant annotation seems not to function in these tests, so do it the other way.
+        try (TenantContext tc = TenantContext.withTenantId(tenantId)) {
+            String integrationKey = "not_verified";
 
-        PagerDutyConfigDTO config = PagerDutyConfigDTO.newBuilder().setIntegrationKey(integrationKey).build();
-        notificationService.postPagerDutyConfig(config);
+            PagerDutyConfigDTO config = PagerDutyConfigDTO.newBuilder().setIntegrationKey(integrationKey).build();
+            notificationService.postPagerDutyConfig(config);
+        }
     }
 
     @Test
     void testProducingAlarmWithConfigSetup() throws NotificationException {
-        setupConfig();
+        setupConfig("opennms-prime");
 
         int id = 1234;
         String tenantId = "opennms-prime";
